Compile MySQL extension as well
The hack for __iob_func being removed from the core runtime, but required by the old mysql we're building against can be simplified a lot due to the `_ReturnAddress` intrinsic available since MSVC 2015.

+++ extensions/mysql/mysql/MyDriver.cpp
@@ -224,7 +224,7 @@ void MyDriver::ShutdownThreadSafety()
 	mysql_thread_end();
 }
 
-unsigned int strncopy(char *dest, const char *src, size_t count)
+size_t strncopy(char *dest, const char *src, size_t count)
 {
 	if (!count)
 	{
